+++ FoodSearch.BusinessLogic.Domain.RestraurantAdmin/Interface/IRestaurantAdminDomain.cs
@@ -22,7 +22,7 @@ namespace FoodSearch.BusinessLogic.Domain.RestraurantAdmin.Interface
         bool EditDishGroup(Guid restaurantId, int dishGroupId, string newGroupName);
         bool DeleteDishGroup(Guid restaurantId, int dishGroupId);
         IEnumerable<Dish> GetDishes(Guid restaurantId);
-        Dish CreateDish(Guid restaurantId, string dishName, int dishGroupId, float price);
+        Dish CreateDish(Guid restaurantId, string dishName, int dishGroupId, decimal price);
         Guid GetUserId(string email);
         Guid GetUserRestaurant(Guid userId);
         IEnumerable<OpeningHour> GetOpeningHours(Guid restaurantId);
