+++ FoodSearch.Data.Mapping/Entities/Dish.cs
@@ -10,6 +10,6 @@ namespace FoodSearch.Data.Mapping.Entities
         public virtual string DishName { get; set; }
         public virtual int DishGroupId { get; set; }
         public virtual DishGroup DishGroup { get; set; }
-        public virtual float Price { get; set; }
+        public virtual decimal Price { get; set; }
     }
 }
