+++ FoodSearch.Data.Mapping/Entities/Restaurant.cs
@@ -10,10 +10,10 @@ namespace FoodSearch.Data.Mapping.Entities
         public virtual int ImageId { get; set; }
         public virtual Image Image { get; set; }
         public virtual string Name { get; set; }
-        public virtual float MinOrderAmount { get; set; }
+        public virtual decimal MinOrderAmount { get; set; }
         public virtual bool IsOpen { get; set; }
         public virtual bool IsDeleted { get; set; }
-        public virtual float DeliveryPrice { get; set; }
-        public virtual float FreeDeliveryFrom { get; set; }
+        public virtual decimal DeliveryPrice { get; set; }
+        public virtual decimal FreeDeliveryFrom { get; set; }
     }
 }
